refactor: 게임 상수 관리 클래스 추가
