+++ src/main/java/racingcar/constants/GameConstants.java
@@ -1,0 +1,19 @@
+package racingcar.constants;
+
+public final class GameConstants {
+
+    public static final int RANDOM_MIN = 0;
+
+    public static final int RANDOM_MAX = 9;
+
+    public static final int MOVE_THRESHOLD = 4;
+
+    public static final int MAX_CAR_NAME_LENGTH = 5;
+
+    public static final int MIN_GAME_TRY_COUNT = 1;
+
+    public static final String COMMA = ",";
+
+    public static final String HYPHEN = "-";
+
+}
