+++ src/main/java/racingcar/constants/MessageType.java
@@ -1,4 +1,4 @@
-package racingcar.view;
+package racingcar.constants;
 
 public enum MessageType {
