+++ src/main/java/racingcar/model/Car.java
@@ -2,12 +2,10 @@ package racingcar.model;
 
 
 import camp.nextstep.edu.missionutils.Randoms;
+import racingcar.constants.GameConstants;
 
 public class Car {
 
-    private static final int RANDOM_MIN = 0;
-    private static final int RANDOM_MAX = 9;
-    private static final int MOVE_THRESHOLD = 4;
     private final String name;
     private int moveCount = 0;
 
@@ -16,8 +14,8 @@ public class Car {
     }
 
     public void moveCar() {
-        int randomNum = Randoms.pickNumberInRange(RANDOM_MIN, RANDOM_MAX);
-        if (randomNum >= MOVE_THRESHOLD) {
+        int randomNum = Randoms.pickNumberInRange(GameConstants.RANDOM_MIN, GameConstants.RANDOM_MAX);
+        if (randomNum >= GameConstants.MOVE_THRESHOLD) {
             moveCount++;
         }
     }
