+++ src/main/java/racingcar/validate/Validation.java
@@ -1,13 +1,11 @@
 package racingcar.validate;
 
-import racingcar.model.ErrorType;
+import racingcar.constants.ErrorType;
+import racingcar.constants.GameConstants;
 
 import java.util.Set;
 
 public class Validation {
-
-    private static final int MAX_CAR_NAME_LENGTH = 5;
-    private static final int MIN_GAME_TRY_COUNT = 1;
 
     public static String validateCreateCar(String input, Set<String> distinctCarsName) {
         String carName = input.trim();
@@ -24,13 +22,13 @@ public class Validation {
     }
 
     private static void validateNameLength(String input) {
-        if(input.trim().length() > MAX_CAR_NAME_LENGTH) {
+        if(input.trim().length() > GameConstants.MAX_CAR_NAME_LENGTH) {
             throw new IllegalArgumentException(ErrorType.CAR_NAME_LENGTH_ERROR.getMessage());
         }
     }
 
     public static void validateGameTryCount(int count) {
-        if(count < MIN_GAME_TRY_COUNT) {
+        if(count < GameConstants.MIN_GAME_TRY_COUNT) {
             throw new IllegalArgumentException(ErrorType.GAME_TRY_COUNT_ERROR.getMessage());
         }
     }
