+++ src/main/java/racingcar/view/InputMessage.java
@@ -1,6 +1,7 @@
 package racingcar.view;
 
 import camp.nextstep.edu.missionutils.Console;
+import racingcar.constants.MessageType;
 
 public class InputMessage {
