+++ src/main/java/racingcar/view/OutputMessage.java
@@ -1,8 +1,9 @@
 package racingcar.view;
 
-public class OutputMessage {
+import racingcar.constants.GameConstants;
+import racingcar.constants.MessageType;
 
-    private static final String HYPHEN = "-";
+public class OutputMessage {
 
     public static void executeResultMessage(String winner) {
         System.out.println(MessageType.RESPONSE_FINAL_WINNER.getMessage() + winner);
@@ -13,7 +14,7 @@ public class OutputMessage {
     }
 
     public static void printHyphenMessage(String carName, int carMoveCount) {
-        System.out.println(carName + " : " + HYPHEN.repeat(carMoveCount));
+        System.out.println(carName + " : " + GameConstants.HYPHEN.repeat(carMoveCount));
     }
 
 }
